pyproject.toml
[tool.poetry]
name = "dataserver"
version = "0.1.0"
description = ""
authors = ["adolf"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
crawlee = {extras = ["beautifulsoup", "playwright"], version = "^0.1.2"}
python-dotenv = "^1.0.1"
firecrawl-py = "^0.0.20"
duckduckgo-search = "^6.2.6"
apify-haystack = "^0.1.3"
mem0ai = "^0.0.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
